Use dependency pinning
For CentOS7, we use dependency pinning in the spec file. Note that we
don't do it for Ubuntu as it will use the dependency versions from the
requirements.txt.

We now use the compatible release operator [1] to define versions in
the requirements.txt.

This commit also removes non relevant warnings when uninstalling
Synergy.

[1]: https://www.python.org/dev/peps/pep-0440/#compatible-release

Change-Id: Icb1dd16d7d8d7246914058a11534bcd099b1f18c