Remove tabulate from the debian dependencies
Synergy depends on the python tabulate package. However, it is not
available as a system package for both CentOS 7 and Ubuntu 14.04.

In this commit we opt for the following solution: don't mark
tabulate as dependency when building synergy system packages, users
will have to install it by hand from PyPI. The synergy package
installation will fail if tabulate is not installed yet.

Change-Id: I76018e1ba27de8ad47b59c85baad2c2d06f03398
Sem-Ver: bugfix