Distribute tabulate as part of Synergy
Synergy depends on the python tabulate package. However, it is not
available as a system package for both CentOS 7 and Ubuntu 14.04.

In this commit we opt for the following solution: don't mark
tabulate as a dependency when building synergy system packages,
instead we provide the tabulate package as part of Synergy.

DocImpact
Change-Id: I76018e1ba27de8ad47b59c85baad2c2d06f03398
Sem-Ver: bugfix