Remove versions for required packages
Requirement versions were set in the first place to limit the
buginess. However, after some time it appears that it made more
harm than good, as discussed in:
https://review.openstack.org/#/c/388003/

In this commit we remove the versions for all the requirements.

In the future, if a specific dependency version if causing some issue
we should reintroduce the version for this dependency to block the
unstable version.

Sem-Ver: bugfix
Change-Id: Ia42509bf2bf9012cc03c8bb25b84f465425e2aad