Change the way we get the version when packaging
We now get the version from the tag on the current commit.
This imply that when we do packaging, the current commit must be tagged
with a version, like "1.3.4".

This change is necessary for the following use case: the codebase is at
a higher version than the version we are packaging. For example, if we
are packaging for version 1.1.0 but the codebase has versions 1.1.0,
1.2.0 and 2.0.0.
With the previous way of packaging, when checking out version 1.1.0 and
packaging, it would say that we are packaging for the version 2.0.0.
This commit changes this behavior and takes the correct version in this
use case.

Change-Id: I37dc60594fe1f570796a057124422e5ddf601269
Sem-Ver: bugfix