Update RPM package dependencies
During a recent installation of Synergy on a standard CentOS 7 cluster
it appeared that python dependencies present in both EPEL and OpenStack
repositories were not installed correctly.
This commit makes the dependencies more explicit and result in a correct
installation procedure.

Note that since the Synergy code has not changed since the last version,
only the package "dist" version is changed.

Change-Id: I1d10fdc7e98788ce83c245aa3606351e358603e1